+++ src/main/kotlin/icu/windea/pls/ep/expression/CwtDataExpressionMatchers.kt
@@ -32,10 +32,10 @@ class BaseCwtDataExpressionMatcher : CwtDataExpressionMatcher {
             }
             configExpression.type == CwtDataTypes.Int -> {
                 //quoted number (e.g. "1") -> ok according to vanilla game files
-                if(expression.type.isIntType() || ParadoxType.resolve(expression.value).isIntType()) {
+                if(expression.type.isIntType() || ParadoxType.resolve(expression.text).isIntType()) {
                     val (min, max) = configExpression.extraValue<Tuple2<Int?, Int?>>() ?: return Result.ExactMatch
                     return Result.LazySimpleMatch p@{
-                        val value = expression.value.toIntOrNull() ?: return@p true
+                        val value = expression.text.toIntOrNull() ?: return@p true
                         (min == null || min <= value) && (max == null || max >= value)
                     }
                 }
@@ -43,10 +43,10 @@ class BaseCwtDataExpressionMatcher : CwtDataExpressionMatcher {
             }
             configExpression.type == CwtDataTypes.Float -> {
                 //quoted number (e.g. "1") -> ok according to vanilla game files
-                if(expression.type.isFloatType() || ParadoxType.resolve(expression.value).isFloatType()) {
+                if(expression.type.isFloatType() || ParadoxType.resolve(expression.text).isFloatType()) {
                     val (min, max) = configExpression.extraValue<Tuple2<Float?, Float?>>() ?: return Result.ExactMatch
                     return Result.LazySimpleMatch p@{
-                        val value = expression.value.toFloatOrNull() ?: return@p true
+                        val value = expression.text.toFloatOrNull() ?: return@p true
                         (min == null || min <= value) && (max == null || max >= value)
                     }
                 }
@@ -65,7 +65,7 @@ class BaseCwtDataExpressionMatcher : CwtDataExpressionMatcher {
                 Result.ofFallback(r)
             }
             configExpression.type == CwtDataTypes.ColorField -> {
-                val r = expression.type.isColorType() && configExpression.value?.let { expression.value.startsWith(it) } != false
+                val r = expression.type.isColorType() && configExpression.value?.let { expression.text.startsWith(it) } != false
                 Result.of(r)
             }
             else -> null
@@ -98,37 +98,37 @@ class CoreCwtDataExpressionMatcher : CwtDataExpressionMatcher {
         return when {
             configExpression.type == CwtDataTypes.PercentageField -> {
                 if(!expression.type.isStringType()) return Result.NotMatch
-                val r = ParadoxTypeHandler.isPercentageField(expression.value)
+                val r = ParadoxTypeHandler.isPercentageField(expression.text)
                 Result.of(r)
             }
             configExpression.type == CwtDataTypes.DateField -> {
                 if(!expression.type.isStringType()) return Result.NotMatch
-                val r = ParadoxTypeHandler.isDateField(expression.value)
+                val r = ParadoxTypeHandler.isDateField(expression.text)
                 Result.of(r)
             }
             configExpression.type == CwtDataTypes.Localisation -> {
                 if(!expression.type.isStringType()) return Result.NotMatch
-                if(!expression.value.isParameterAwareIdentifier('.', '-', '\'')) return Result.NotMatch
+                if(!expression.text.isParameterAwareIdentifier('.', '-', '\'')) return Result.NotMatch
                 if(expression.isParameterized()) return Result.ParameterizedMatch
                 CwtConfigMatcher.Impls.getLocalisationMatchResult(element, expression, project)
             }
             configExpression.type == CwtDataTypes.SyncedLocalisation -> {
                 if(!expression.type.isStringType()) return Result.NotMatch
-                if(!expression.value.isParameterAwareIdentifier('.', '-', '\'')) return Result.NotMatch
+                if(!expression.text.isParameterAwareIdentifier('.', '-', '\'')) return Result.NotMatch
                 if(expression.isParameterized()) return Result.ParameterizedMatch
                 CwtConfigMatcher.Impls.getSyncedLocalisationMatchResult(element, expression, project)
             }
             configExpression.type == CwtDataTypes.InlineLocalisation -> {
                 if(!expression.type.isStringType()) return Result.NotMatch
                 if(expression.quoted) return Result.FallbackMatch //"quoted_string" -> any string
-                if(!expression.value.isParameterAwareIdentifier('.', '-', '\'')) return Result.NotMatch
+                if(!expression.text.isParameterAwareIdentifier('.', '-', '\'')) return Result.NotMatch
                 if(expression.isParameterized()) return Result.ParameterizedMatch
                 CwtConfigMatcher.Impls.getSyncedLocalisationMatchResult(element, expression, project)
             }
             configExpression.type == CwtDataTypes.Definition -> {
                 //can be a integer here (e.g., for <technology_tier>)
                 if(!expression.type.isStringType() && expression.type != ParadoxType.Int) return Result.NotMatch
-                if(!expression.value.isParameterAwareIdentifier('.')) return Result.NotMatch
+                if(!expression.text.isParameterAwareIdentifier('.')) return Result.NotMatch
                 if(expression.isParameterized()) return Result.ParameterizedMatch
                 CwtConfigMatcher.Impls.getDefinitionMatchResult(element, expression, configExpression, project)
             }
@@ -144,7 +144,7 @@ class CoreCwtDataExpressionMatcher : CwtDataExpressionMatcher {
             configExpression.type == CwtDataTypes.EnumValue -> {
                 if(expression.type.isBlockLikeType()) return Result.NotMatch
                 if(expression.isParameterized()) return Result.ParameterizedMatch
-                val name = expression.value
+                val name = expression.text
                 val enumName = configExpression.value ?: return Result.NotMatch //invalid cwt config
                 //match simple enums
                 val enumConfig = configGroup.enums[enumName]
@@ -156,7 +156,7 @@ class CoreCwtDataExpressionMatcher : CwtDataExpressionMatcher {
                 val complexEnumConfig = configGroup.complexEnums[enumName]
                 if(complexEnumConfig != null) {
                     //complexEnumValue的值必须合法
-                    if(ParadoxComplexEnumValueHandler.getName(expression.value) == null) return Result.NotMatch
+                    if(ParadoxComplexEnumValueHandler.getName(expression.text) == null) return Result.NotMatch
                     return CwtConfigMatcher.Impls.getComplexEnumValueMatchResult(element, name, enumName, complexEnumConfig, project)
                 }
                 Result.NotMatch
@@ -165,7 +165,7 @@ class CoreCwtDataExpressionMatcher : CwtDataExpressionMatcher {
                 if(expression.type.isBlockLikeType()) return Result.NotMatch
                 if(expression.isParameterized()) return Result.ParameterizedMatch
                 //dynamicValue的值必须合法
-                val name = ParadoxDynamicValueHandler.getName(expression.value)
+                val name = ParadoxDynamicValueHandler.getName(expression.text)
                 if(name == null) return Result.NotMatch
                 val dynamicValueType = configExpression.value
                 if(dynamicValueType == null) return Result.NotMatch
@@ -179,42 +179,42 @@ class CoreCwtDataExpressionMatcher : CwtDataExpressionMatcher {
             configExpression.type in CwtDataTypeGroups.ValueField -> {
                 //也可以是数字，注意：用括号括起的数字（作为scalar）也匹配这个规则
                 if(configExpression.type == CwtDataTypes.ValueField) {
-                    if(expression.type.isFloatType() || ParadoxType.resolve(expression.value).isFloatType()) return Result.ExactMatch
+                    if(expression.type.isFloatType() || ParadoxType.resolve(expression.text).isFloatType()) return Result.ExactMatch
                 } else if(configExpression.type == CwtDataTypes.IntValueField) {
-                    if(expression.type.isIntType() || ParadoxType.resolve(expression.value).isIntType()) return Result.ExactMatch
+                    if(expression.type.isIntType() || ParadoxType.resolve(expression.text).isIntType()) return Result.ExactMatch
                 }
                 if(!expression.type.isStringType()) return Result.NotMatch
                 if(expression.isParameterized()) return Result.ParameterizedMatch
-                val textRange = TextRange.create(0, expression.value.length)
-                val valueFieldExpression = ParadoxValueFieldExpression.resolve(expression.value, textRange, configGroup)
+                val textRange = TextRange.create(0, expression.text.length)
+                val valueFieldExpression = ParadoxValueFieldExpression.resolve(expression.text, textRange, configGroup)
                 val r = valueFieldExpression != null
                 Result.of(r)
             }
             configExpression.type in CwtDataTypeGroups.VariableField -> {
                 //也可以是数字，注意：用括号括起的数字（作为scalar）也匹配这个规则
                 if(configExpression.type == CwtDataTypes.VariableField) {
-                    if(expression.type.isFloatType() || ParadoxType.resolve(expression.value).isFloatType()) return Result.ExactMatch
+                    if(expression.type.isFloatType() || ParadoxType.resolve(expression.text).isFloatType()) return Result.ExactMatch
                 } else if(configExpression.type == CwtDataTypes.IntVariableField) {
-                    if(expression.type.isIntType() || ParadoxType.resolve(expression.value).isIntType()) return Result.ExactMatch
+                    if(expression.type.isIntType() || ParadoxType.resolve(expression.text).isIntType()) return Result.ExactMatch
                 }
                 if(!expression.type.isStringType()) return Result.NotMatch
                 if(expression.isParameterized()) return Result.ParameterizedMatch
-                val textRange = TextRange.create(0, expression.value.length)
-                val variableFieldExpression = ParadoxVariableFieldExpression.resolve(expression.value, textRange, configGroup)
+                val textRange = TextRange.create(0, expression.text.length)
+                val variableFieldExpression = ParadoxVariableFieldExpression.resolve(expression.text, textRange, configGroup)
                 val r = variableFieldExpression != null
                 Result.of(r)
             }
             configExpression.type in CwtDataTypeGroups.DatabaseObject -> {
                 if(!expression.type.isStringType()) return Result.NotMatch
                 if(expression.isParameterized()) return Result.ParameterizedMatch
-                val textRange = TextRange.create(0, expression.value.length)
-                val databaseObjectExpression = ParadoxDatabaseObjectExpression.resolve(expression.value, textRange, configGroup)
+                val textRange = TextRange.create(0, expression.text.length)
+                val databaseObjectExpression = ParadoxDatabaseObjectExpression.resolve(expression.text, textRange, configGroup)
                 val r = databaseObjectExpression != null
                 Result.of(r)
             }
             configExpression.type == CwtDataTypes.Modifier -> {
                 if(!expression.type.isStringType()) return Result.NotMatch
-                if(!expression.value.isParameterAwareIdentifier()) return Result.NotMatch
+                if(!expression.text.isParameterAwareIdentifier()) return Result.NotMatch
                 if(expression.isParameterized()) return Result.ParameterizedMatch
                 CwtConfigMatcher.Impls.getModifierMatchResult(element, expression, configGroup)
             }
@@ -225,14 +225,14 @@ class CoreCwtDataExpressionMatcher : CwtDataExpressionMatcher {
                 if(!expression.type.isStringLikeType()) return Result.NotMatch
                 if(expression.isParameterized()) return Result.ParameterizedMatch
                 val aliasName = configExpression.value ?: return Result.NotMatch
-                val aliasSubName = ParadoxExpressionHandler.getAliasSubName(element, expression.value, expression.quoted, aliasName, configGroup, options) ?: return Result.NotMatch
+                val aliasSubName = ParadoxExpressionHandler.getAliasSubName(element, expression.text, expression.quoted, aliasName, configGroup, options) ?: return Result.NotMatch
                 CwtConfigMatcher.matches(element, expression, CwtDataExpression.resolve(aliasSubName, true), null, configGroup)
             }
             configExpression.type == CwtDataTypes.AliasName -> {
                 if(!expression.type.isStringLikeType()) return Result.NotMatch
                 if(expression.isParameterized()) return Result.ParameterizedMatch
                 val aliasName = configExpression.value ?: return Result.NotMatch
-                val aliasSubName = ParadoxExpressionHandler.getAliasSubName(element, expression.value, expression.quoted, aliasName, configGroup, options) ?: return Result.NotMatch
+                val aliasSubName = ParadoxExpressionHandler.getAliasSubName(element, expression.text, expression.quoted, aliasName, configGroup, options) ?: return Result.NotMatch
                 CwtConfigMatcher.matches(element, expression, CwtDataExpression.resolve(aliasSubName, true), null, configGroup)
             }
             configExpression.type == CwtDataTypes.AliasMatchLeft -> {
@@ -244,7 +244,7 @@ class CoreCwtDataExpressionMatcher : CwtDataExpressionMatcher {
             configExpression.type == CwtDataTypes.Parameter -> {
                 //匹配参数名（即使对应的定义声明中不存在对应名字的参数，也可以匹配）
                 if(!expression.type.isStringLikeType()) return Result.NotMatch
-                if(!expression.value.isParameterAwareIdentifier()) return Result.NotMatch
+                if(!expression.text.isParameterAwareIdentifier()) return Result.NotMatch
                 Result.ExactMatch
             }
             configExpression.type == CwtDataTypes.ParameterValue -> {
@@ -255,7 +255,7 @@ class CoreCwtDataExpressionMatcher : CwtDataExpressionMatcher {
             configExpression.type == CwtDataTypes.LocalisationParameter -> {
                 //匹配本地化参数名（即使对应的定义声明中不存在对应名字的参数，也可以匹配）
                 if(!expression.type.isStringLikeType()) return Result.NotMatch
-                if(!expression.value.isParameterAwareIdentifier('.', '-', '\'')) return Result.NotMatch
+                if(!expression.text.isParameterAwareIdentifier('.', '-', '\'')) return Result.NotMatch
                 Result.ExactMatch
             }
             configExpression.type == CwtDataTypes.ShaderEffect -> {
@@ -272,7 +272,7 @@ class CoreCwtDataExpressionMatcher : CwtDataExpressionMatcher {
             }
             configExpression.type == CwtDataTypes.TechnologyWithLevel -> {
                 if(!expression.type.isStringType()) return Result.NotMatch
-                if(expression.value.length > 1 && expression.value.indexOf('@') >= 1) return Result.ExactMatch
+                if(expression.text.length > 1 && expression.text.indexOf('@') >= 1) return Result.ExactMatch
                 if(expression.isParameterized()) return Result.ParameterizedMatch
                 Result.NotMatch
             }
@@ -287,7 +287,7 @@ class ConstantCwtDataExpressionMatcher : CwtDataExpressionMatcher, CwtConfigPatt
             val value = configExpression.value ?: return Result.NotMatch
             if(!configExpression.isKey) {
                 //常量的值也可能是yes/no
-                val text = expression.value
+                val text = expression.text
                 if((value == "yes" || value == "no") && text.isLeftQuoted()) return Result.NotMatch
             }
             //这里也用来匹配空字符串
@@ -315,7 +315,7 @@ class AntExpressionCwtDataExpressionMatcher : CwtDataExpressionMatcher, CwtConfi
         if(configExpression.type == CwtDataTypes.AntExpression) {
             val pattern = configExpression.value ?: return Result.NotMatch
             val ignoreCase = configExpression.extraValue?.castOrNull<Boolean>() ?: false
-            val r = expression.value.matchesAntPattern(pattern, ignoreCase)
+            val r = expression.text.matchesAntPattern(pattern, ignoreCase)
             return Result.of(r)
         }
         return null
@@ -327,7 +327,7 @@ class RegexCwtDataExpressionMatcher : CwtDataExpressionMatcher, CwtConfigPattern
         if(configExpression.type == CwtDataTypes.Regex) {
             val pattern = configExpression.value ?: return Result.NotMatch
             val ignoreCase = configExpression.extraValue?.castOrNull<Boolean>() ?: false
-            val r = expression.value.matchesRegex(pattern, ignoreCase)
+            val r = expression.text.matchesRegex(pattern, ignoreCase)
             return Result.of(r)
         }
         return null