+++ src/main/kotlin/icu/windea/pls/model/expression/ParadoxDataExpression.kt
@@ -14,39 +14,39 @@ import java.util.*
  * 数据表达式，对应脚本语言中的某处键/值。
  */
 class ParadoxDataExpression private constructor(
-    val value: String,
+    val text: String,
     val type: ParadoxType,
     val quoted: Boolean,
     val isKey: Boolean?
 ) {
-    private val regexWhenIsParameterized by lazy { value.toRegexWhenIsParameterized() }
+    private val regexWhenIsParameterized by lazy { text.toRegexWhenIsParameterized() }
     
     fun isParameterized(): Boolean {
-        return type == ParadoxType.String && value.isParameterized()
+        return type == ParadoxType.String && text.isParameterized()
     }
     
     fun isFullParameterized(): Boolean {
-        return type == ParadoxType.String && value.isFullParameterized()
+        return type == ParadoxType.String && text.isFullParameterized()
     }
     
     fun matchesConstant(v: String): Boolean {
-        if(value.isParameterized()) {
+        if(text.isParameterized()) {
             //兼容带参数的情况（此时先转化为正则表达式，再进行匹配）
             return regexWhenIsParameterized.matches(v)
         }
-        return value.equals(v, true) //忽略大小写
+        return text.equals(v, true) //忽略大小写
     }
     
     override fun equals(other: Any?): Boolean {
-        return other is ParadoxDataExpression && (value == other.value && quoted == other.quoted)
+        return other is ParadoxDataExpression && (text == other.text && quoted == other.quoted)
     }
     
     override fun hashCode(): Int {
-        return Objects.hash(value, type)
+        return Objects.hash(text, type)
     }
     
     override fun toString(): String {
-        return value
+        return text
     }
     
     companion object Resolver {